reubicación y estandarización
