+++ modules/core/test/fixtures/SecurityFixtures.scala
@@ -2,7 +2,7 @@ package fixtures
 
 import security.*
 import types.Permission
-import user.{Role, UserStatus}
+import user.Role
 
 object SecurityFixtures:
 
@@ -18,32 +18,6 @@ object SecurityFixtures:
 
   val requestToken = RequestToken("test-request-token-123")
 
-  val ldapUser = LdapUser(
-    userName = "testuser",
-    firstName = "Test",
-    lastName = "User",
-    email = "test@example.com",
-    roles = Seq("geneticist", "clerk"),
-    geneMapperId = "TLAB",
-    phone1 = "1234567890",
-    phone2 = Some("0987654321"),
-    status = UserStatus.active,
-    encryptedPublicKey = Array.emptyByteArray,
-    encryptedPrivateKey = Array.emptyByteArray,
-    encryptrdTotpSecret = totpSecret.getBytes("UTF-8"),
-    superuser = false
-  )
-
-  val blockedLdapUser = ldapUser.copy(
-    userName = "blockeduser",
-    status = UserStatus.blocked
-  )
-
-  val pendingLdapUser = ldapUser.copy(
-    userName = "pendinguser",
-    status = UserStatus.pending
-  )
-
   val rolePermissionMap: Map[String, Set[Permission]] = Map(
     "geneticist" -> Set(Permission.DNA_PROFILE_CRUD, Permission.MATCHES_MANAGER),
     "clerk" -> Set(Permission.PROFILE_DATA_CRUD, Permission.USER_CRUD)
@@ -56,27 +30,13 @@ object SecurityFixtures:
     Permission.USER_CRUD
   )
 
-  val user = User(
-    id = "testuser",
-    firstName = "Test",
-    lastName = "User",
-    email = "test@example.com",
-    geneMapperId = "TLAB",
-    roles = Seq("geneticist", "clerk"),
-    permissions = userPermissions,
-    status = UserStatus.active,
-    phone1 = "1234567890",
-    phone2 = Some("0987654321"),
-    superuser = false
-  )
-
   val userCredentials = UserCredentials(
     publicKey = Array.emptyByteArray,
     privateKey = Array.emptyByteArray,
     totpSecret = totpSecret
   )
 
-  val fullUser = FullUser(user, userCredentials, authPair)
+  val fullUser = FullUser(UserFixtures.user, userCredentials, authPair)
 
   val geneticistRole = Role("geneticist", "Geneticist", Set(Permission.DNA_PROFILE_CRUD, Permission.MATCHES_MANAGER))
   val clerkRole = Role("clerk", "Clerk", Set(Permission.PROFILE_DATA_CRUD, Permission.USER_CRUD))
